+++ src/app/orders-table/orders-table.component.ts
@@ -98,18 +98,22 @@ export class OrdersTableComponent {
   }
 
   deleteRecord(event) {
-    console.log(event.data);
-    this.http.delete('http://localhost:4321/deleteOrder/' + event.data.id).subscribe(
-      res => {
-        console.log(res);
-        event.confirm.resolve(event.source.data);
-      },
-      (err: HttpErrorResponse) => {
-        if (err.error instanceof Error) {
-          console.log('Client-side error occured.');
-        } else {
-          console.log('Server-side error occured.');
-        }
-      });
+    if (window.confirm('Are you sure you want to delete?')) {
+      console.log(event.data);
+      this.http.delete('http://localhost:4321/deleteOrder/' + event.data.id).subscribe(
+        res => {
+          console.log(res);
+          event.confirm.resolve(event.source.data);
+        },
+        (err: HttpErrorResponse) => {
+          if (err.error instanceof Error) {
+            console.log('Client-side error occured.');
+          } else {
+            console.log('Server-side error occured.');
+          }
+        });
+    } else {
+      event.confirm.reject();
+    }
   }
 }
